Implement search bar functionality and refactor media grid creation
